2Q-PoC/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
aspose-words==24.9.0
asttokens==2.4.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-search-documents==11.5.2
azure-storage-blob==12.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
doc2docx==0.2.4
docx2pdf==0.1.8
executing==2.1.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
jiter==0.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
nest-asyncio==1.6.0
numpy==2.1.1
oauthlib==3.2.2
openai==1.47.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.6
plum-dispatch==1.7.4
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==305
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
Spire.Doc==12.7.1
stack-data==0.6.3
starlette==0.41.3
SudachiDict-core==20240716
SudachiPy==0.6.8
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
unoconv==0.9.0
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
